removed the need for simhub; made custom python client communication via UDP; made custom ac lua app for extensive telemetry; added turn signal support, handbrake support + lighting support & sped up polling rate

+++ src/commands/fuel.cpp
@@ -6,7 +6,7 @@
 
 
 void sendFuelLevel(int percent){
-    uint16_t sensor = percent * 160;
+    uint16_t sensor = static_cast<uint16_t>(percent * 160);
 
 
     CAN.beginPacket(0x349);
